+++ 0x1E-search_algorithms/search_algos.h
@@ -3,13 +3,12 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include "search_algos.h"
 
 int linear_search(int *array, size_t size, int value);
 int binary_search(int *array, size_t size, int value);
 int jump_search(int *array, size_t size, int value);
 
 
-void printer(int *array, int i, int n);
+void printer(int *array, size_t i, size_t n);
 
 #endif
